multiline support for textinput with ListView placement compensation
set `multiline` to true as a property to `GiftedMessenger` to enable. set `maxInputHeight` to specify the maximum height the input can grow too. If you need the change event for when it changes height, use `onTextInputHeightChanged`. You can pass additional props to the input now with `textInputProps`. 

The autogrowing part was easy. What was very time-consuming was getting the positioning compensation for the ListView so the correct part was always placed properly. I had to animate the textinput separately (well, its container). So similar to the ListView container, I animate its height. They are both done in parallel, but could likely be optimized to interpolate. The reason it had to be done separately--and believe me, I tried to use the initial mechanism to move the input up--is because when the placement of the textinput is dependent on the height of the ListVIew, then you can't move the ListView (in relation to the TextInput) without also moving the TextInput. I spent a long time trying to do that via various ways and things just got tangled, so I did what I probably should have originally did which was isolate the TextInput so it moves in an unrelated fashion. Absolute positioning and a `bottom` property are used to have the textinput stick to the bottom and grow up from the bottom. Aside from the listview placement compensation, that's its secret sauce.