fix(Bubble): re-enable generics on message prop
PR #2609 introduced a regression in the types by migrating the Bubble
component from a class component to using React.FC, which does not support generics.

This commit replaces React.FC with generics and a JSX.Element return type on the
Bubble component, allowing to keep the function component and bringing back the
type argument.